refactored the code so that we have CategoryGraph separate component
