+++ src/components/navbar/navbar.module.css
@@ -1,12 +1,3 @@
-@import url('https://fonts.cdnfonts.com/css/euclid-circular-b');
-@font-face {
-  font-family: 'Euclid Circular B';
-  src: url('/fonts/Euclid\ Circular\ B\ Bold.ttf Circular B Bold Italic.ttf') format('ttf');
-     
-  /* Add more source formats if needed (TTF, OTF, etc.) */
-  font-weight: normal;
-  font-style: normal;
-}
 
 .image{
     width: 5rem;
@@ -24,7 +15,7 @@
 }
 .title{
   /* font-family: 'sans-serif'; */
-  font-family: 'Euclid Circular B';
+  font-family: 'Poppins';
 }
 
 .desc p{
